1.dubbo对内开发 2.cpc计算开发

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/report/MktCampaingReportServiceImpl.java
@@ -148,7 +148,7 @@ public class MktCampaingReportServiceImpl implements MktCampaingReportService {
             int count = 0;
             for (MktCampaignDO mktCampaignDO : mktCampaignDOInList) {
                 // "不活跃活动"活动判断地市
-                if (sysArea.getAreaId().equals(mktCampaignDO.getLanId())) {
+                if (sysArea.getAreaId().equals(mktCampaignDO.getLanId().intValue())) {
                     count++;
                 }
             }
